+++ src/chrome/backgroundscript.js
@@ -1,6 +1,6 @@
 /*
  * Copyright Adam Pritchard 2016
- * MIT License : http://adampritchard.mit-license.org/
+ * MIT License : https://adampritchard.mit-license.org/
  */
 
 "use strict";
@@ -8,20 +8,59 @@
   marked:false, hljs:false, Utils:false, CommonLogic:false */
 /*jshint devel:true, browser:true*/
 
-/*
- * Chrome background script.
- */
-
-// On each load, check if we should show the options/changelist page.
-function onLoad() {
-  // This timeout is a dirty hack to fix bug #119: "Markdown Here Upgrade
-  // Notification every time I open Chrome". That issue on Github for details.
-  // https://github.com/adam-p/markdown-here/issues/119
-  window.setTimeout(upgradeCheck, 30000);
+if (typeof browser === "undefined") {
+  // Chrome does not support the browser namespace yet.
+  // See https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/manifest.json/background
+  globalThis.browser = chrome;
 }
 
-// In the interest of improved browser load performace, call `onLoad` after a tick.
-window.addEventListener('load', Utils.nextTickFn(onLoad), false);
+// We supply a #hash to the background page, so that we know when we're
+// loaded via `background.page` (manifest V2 and Firefox manifest V3) vs
+// `background.service_worker` (manifest V3 in Chrome and Safari).
+var backgroundPage = !!location.hash;
+
+if (!backgroundPage) {
+  // When loaded via a background page, the support scripts are already
+  // present. When loaded via a service worker, we need to import them.
+  // (`importScripts` is only available in service workers.)
+  importScripts('../common/utils.js');
+  importScripts('../common/common-logic.js');
+  importScripts('../common/marked.js');
+  importScripts('../common/highlightjs/highlight.js');
+  importScripts('../common/markdown-render.js');
+  importScripts('../common/options-store.js');
+}
+
+// Note that this file is both the script for a background page _and_ for a service
+// worker. The way these things work are quite different, and we must be cognizant of that
+// while writing this file.
+//
+// The key difference is that a background page is loaded once per browser session; a
+// service worker is loaded when extension-related events occur, and then is torn down
+// after 30 seconds of inactivity (with lifecycle caveats). This means that we can't rely
+// on global variables to store state, and we must be mindful about how we handle
+// messages.
+
+// For the background page, this listener is added once and remains active for the browser
+// session; for the service worker, this listener is added every time the service worker
+// is loaded, and is torn down when the service worker is torn down.
+chrome.runtime.onInstalled.addListener((details) => {
+  if (details.reason !== 'install' && details.reason !== 'update') {
+    return;
+  }
+
+  // Create the context menu that will signal our main code.
+  // This must be called only once, when installed or updated, so we do it here.
+  chrome.contextMenus.create({
+    id: 'markdown-here-context-menu',
+    contexts: ['editable'],
+    title: Utils.getMessage('context_menu_item')
+  });
+
+  // Note: If we find that the upgrade info page opens too often, we may
+  // need to add delays. See: https://github.com/adam-p/markdown-here/issues/119
+  upgradeCheck();
+});
 
 function upgradeCheck() {
   OptionsStore.get(function(options) {
@@ -35,7 +74,7 @@ function upgradeCheck() {
       OptionsStore.set({ 'last-version': appManifest.version }, function() {
         // This is the very first time the extensions has been run, so show the
         // options page.
-        chrome.tabs.create({ url: chrome.extension.getURL(optionsURL) });
+        chrome.tabs.create({ url: chrome.runtime.getURL(optionsURL) });
       });
     }
     else if (options['last-version'] !== appManifest.version) {
@@ -45,23 +84,20 @@ function upgradeCheck() {
         // The extension has been newly updated
         optionsURL += '?prevVer=' + options['last-version'];
 
-        showUpgradeNotification(chrome.extension.getURL(optionsURL));
+        showUpgradeNotification(chrome.runtime.getURL(optionsURL));
       });
     }
   });
 }
 
-// Create the context menu that will signal our main code.
-chrome.contextMenus.create({
-  contexts: ['editable'],
-  title: Utils.getMessage('context_menu_item'),
-  onclick: function(info, tab) {
-    chrome.tabs.sendMessage(tab.id, {action: 'context-click'});
-  }
+// Handle context menu clicks.
+chrome.contextMenus.onClicked.addListener(function(info, tab) {
+  chrome.tabs.sendMessage(tab.id, {action: 'context-click'});
 });
 
-// Handle rendering requests from the content script.
-// See the comment in markdown-render.js for why we do this.
+// Handle rendering requests from the content script. Note that incoming messages will
+// revive the service worker, then process the message, then tear down the service worker.
+// See the comment in markdown-render.js for why we use these requests.
 chrome.runtime.onMessage.addListener(function(request, sender, responseCallback) {
   // The content script can load in a not-real tab (like the search box), which
   // has an invalid `sender.tab` value. We should just ignore these pages.
@@ -89,11 +125,11 @@ chrome.runtime.onMessage.addListener(function(request, sender, responseCallback)
   }
   else if (request.action === 'show-toggle-button') {
     if (request.show) {
-      chrome.browserAction.enable(sender.tab.id);
-      chrome.browserAction.setTitle({
+      chrome.action.enable(sender.tab.id);
+      chrome.action.setTitle({
         title: Utils.getMessage('toggle_button_tooltip'),
         tabId: sender.tab.id });
-      chrome.browserAction.setIcon({
+      chrome.action.setIcon({
         path: {
           "16": Utils.getLocalURL('/common/images/icon16-button-monochrome.png'),
           "19": Utils.getLocalURL('/common/images/icon19-button-monochrome.png'),
@@ -105,11 +141,11 @@ chrome.runtime.onMessage.addListener(function(request, sender, responseCallback)
       return false;
     }
     else {
-      chrome.browserAction.disable(sender.tab.id);
-      chrome.browserAction.setTitle({
+      chrome.action.disable(sender.tab.id);
+      chrome.action.setTitle({
         title: Utils.getMessage('toggle_button_tooltip_disabled'),
         tabId: sender.tab.id });
-      chrome.browserAction.setIcon({
+      chrome.action.setIcon({
         path: {
           "16": Utils.getLocalURL('/common/images/icon16-button-disabled.png'),
           "19": Utils.getLocalURL('/common/images/icon19-button-disabled.png'),
@@ -148,7 +184,7 @@ chrome.runtime.onMessage.addListener(function(request, sender, responseCallback)
 });
 
 // Add the browserAction (the button in the browser toolbar) listener.
-chrome.browserAction.onClicked.addListener(function(tab) {
+chrome.action.onClicked.addListener(function(tab) {
   chrome.tabs.sendMessage(tab.id, {action: 'button-click', });
 });
 
@@ -191,6 +227,10 @@ function showUpgradeNotification(optionsURL) {
       });
     };
 
+    // TODO: This interval won't keep the service worker alive, so if a content script
+    // doesn't reload in about 30 seconds, we'll lose the interval and the notification
+    // won't show.
+    // Maybe use the Alarms API? Maybe restructure this so that it's less hacky?
     showUpgradeNotificationInterval = setInterval(askTabsToShowNotification, 5000);
   });
 }