+++ src/common/CHANGES.md
@@ -1,10 +1,36 @@
 Change Log
 ==========
 
+2024-xx-yy: v?
+-------------------
+
+* Replaced links to our Google Group with links to our [GitHub Discussions](https://github.com/adam-p/markdown-here/discussions).
+
+* Removed support for Postbox, which stopped allowing third-party extensions a while ago. Also removed support for "Inbox by Google", which stopped existing, like, a long time ago.
+
+2024-11-09: v2.14.2
+-------------------
+
+* Fixed [issue #261](https://github.com/adam-p/markdown-here/issues/261): Switched from using the Google Chart API to CodeCogs for the the TeX renderer. Google shut down the API we were using, so it just didn't work anymore.
+  - Take a look at [the wiki](https://github.com/adam-p/markdown-here/wiki/Tips-and-Tricks#tex) for information on using other rendering services.
+  - Some day maybe we'll figure out how to [render locally](https://github.com/adam-p/markdown-here/issues/874).
+
+2024-10-17: v2.14.1
+-------------------
+
+* Fixed error on the options page.
+
+2024-10-03: v2.14.0
+-------------------
+
+* Fixed [bug #722](https://github.com/adam-p/markdown-here/issues/722): Added support for WebExtensions manifest V3. (If you've seen warnings about the extension lately, this was why.)
+  - Thanks to [Andrew M. MacFie](https://github.com/amacfie) and [Alexander Popov](https://github.com/AlexWayfer).
+* Fixed [bug #865](https://github.com/adam-p/markdown-here/issues/865): Add Chrome Store privacy information.
+
 2018-09-30: v2.13.4
 --------------------
 
-* Fixed [bug #524] and [bug #526]: Due to v2.13.3 fix, Markdown Here didn't work in Thunderbird with a non-English UI.
+* Fixed [bug #524](https://github.com/adam-p/markdown-here/issues/524) and [bug #526](https://github.com/adam-p/markdown-here/issues/526): Due to v2.13.3 fix, Markdown Here didn't work in Thunderbird with a non-English UI.
   - Thanks to [KSR-Yasuda](https://github.com/KSR-Yasuda), [ensleep](https://github.com/ensleep), [Pedro Silva](https://github.com/pmanu93), [Christophe Meyer](https://github.com/stombi), [littdky](https://github.com/littdky), [Michael Lashkevich](https://github.com/lashkevi), [morsedl](https://github.com/morsedl).
 
 
@@ -39,10 +65,10 @@ Change Log
 2015-09-07: v2.12.0
 --------------------
 
-* **New logo!** Created by [Austin Anderson](http://protractor.ninja/) and chosen by the Markdown Here users, our great new logo is finally ready for action.
+* **New logo!** Created by [Austin Anderson](https://www.angerson.org/) and chosen by the Markdown Here users, our great new logo is finally ready for action.
   - Much thanks to Austin for his work polishing and tweaking the 60 variants of the logo. He did a lot of work getting things just right.
   - Thanks to everyone else who submitted a design: [Rafe Goldberg](http://rgd2.co/), [Jack Reis](http://www.ux-jack.com/), [Enrique Esquinas](https://twitter.com/esquinas), [Leon Wilmanns](http://leon-wilmanns.de/). All of their designs were great and I would have been happy for any of them to win.
-  - Seriously, if you haven't seen the other designs, [go check them out](http://markdown-here.com/logo.html). And then check out their portfolios, bookmark them, and contact them the next time you need design work done.
+  - Seriously, if you haven't seen the other designs, [go check them out](https://markdown-here.com/logo.html). And then check out their portfolios, bookmark them, and contact them the next time you need design work done.
   - Thanks to all of you who took the time to vote for the new logo.
   - If you are unhappy with the change to a monochrome browser toolbar and context menu button, [\+1 this issue](https://github.com/adam-p/markdown-here/issues/302).
 
@@ -369,7 +395,7 @@ Added support for [**Opera**](http://www.opera.com)! Get it [here](https://addon
 2013-05-19: v2.8.0
 ------------------
 
-* **Markdown Here now supports Safari!** [Get it here.](http://markdown-here.com/get.html)
+* **Markdown Here now supports Safari!** [Get it here.](https://markdown-here.com/get.html)
   * Thanks to users [unscriptable](https://github.com/unscriptable) and [martinsvalin](https://github.com/martinsvalin) for [requesting it](https://github.com/adam-p/markdown-here/issues/38).
 
 * Fixed bug in Firefox v23 (current Aurora): Options page wasn't working. ([See details.](https://github.com/adam-p/markdown-here/commit/c20b7e4841f325bed3201ea9a98b3f6c986cf8cc))